Update README.md
I saw the word invocation a lot in this text, which I took to be translated from вызов. I changed it to "command' because I've never really heard this word used in this context. If this is an esoteric usage that I am unaware of, please let me know. Conversely, using the the verb form 'to invoke' when referring to a command is totally appropriate. This is a really well written text.